Added comment and moved SessionViewModel
