+++ src/main/java/retrieve_timetable_use_case/application_business/RetrieveTimetableView.java
@@ -1,6 +1,6 @@
 package retrieve_timetable_use_case.application_business;
 
-import screens.SessionViewModel;
+import display_timetable_use_case.frameworks_and_drivers.SessionViewModel;
 
 /**
  * A view used in the RetrieveTimetable use case that holds and can update its session.
